feat: Add D3.js library and code for drawing a complete graph
This commit adds the D3.js library and the necessary code to draw a complete graph on the web page. The graph is generated based on the number of vertices specified. This feature will be used to visualize the output of the Prim and Kruskal algorithms.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature implementation.